#157 fix the bug that the history_log file is created again when the scan is completed

+++ src/logjam-ingest/ingest.py
@@ -189,14 +189,15 @@ def search_case_directory(scan_obj, search_dir, case_num):
     send them to a running Elastissearch service via the Elastisearch object `es_obj`.
     """
     child_scan = incremental.WorkerScan(search_dir, scan_obj.history_dir, scan_obj.scratch_dir, str(case_num) + ".txt", str(case_num) + "-log.txt", scan_obj.safe_time)
-    es_obj = get_es_connection()
-    recursive_search(child_scan, search_dir, es_obj, case_num)
-    global graceful_abort
-    if graceful_abort:
-        child_scan.premature_exit()
-    else:
-        child_scan.complete_scan()
-        unzip.delete_file(child_scan.history_log_file)
+    if child_scan:
+        es_obj = get_es_connection()
+        recursive_search(child_scan, search_dir, es_obj, case_num)
+        global graceful_abort
+        if graceful_abort:
+            child_scan.premature_exit()
+        else:
+            child_scan.complete_scan()
+            unzip.delete_file(child_scan.history_log_file)
     return
